reftest: R1 render_to_png rasterizer core
Standalone tool rasterizing an HTML file to PNG via the test container
(canvas_ity software raster + Ahem/Terminus fonts + lodepng). A reftest compares
test vs ref PNGs from this same tool, so font/AA differences cancel out.

- render_to_png.cpp: createFromString -> render(width) -> draw(canvas) -> PNG;
  RTP_DEBUG / RTP_SELFTEST env switches for diagnostics
- build_render_to_png.cmd: one-shot cl build (c++20 for iso646 alt tokens;
  gumbo private + visualc includes)
- test container: unqualified std names (string/vector/min/max/optional/swap)
  and pixel_t<int> comparisons fixed to compile against this litehtml version
- fonts/ahem.ttf: the WPT Ahem reftest font

Verified: block backgrounds and grid placement rasterize to correct pixels.

+++ containers/test/Bitmap.cpp
@@ -74,9 +74,9 @@ void Bitmap::draw_rect(int x, int y, int _width, int _height, color color)
 
 void Bitmap::fill_rect(rect rect, color color)
 {
-    for(int y = (int) rect.top(); y < rect.bottom(); y++)
+    for(int y = (int) rect.top(); y < (int) rect.bottom(); y++)
     {
-        for(int x = (int) rect.left(); x < rect.right(); x++)
+        for(int x = (int) rect.left(); x < (int) rect.right(); x++)
         {
             set_pixel(x, y, color);
         }
